Enhance Houdini and Nuke JSON configurations with environment variables. Update PrefsUtils to set environment variables from JSON and improve dark mode handling in SystemUtils.

+++ startup/utils/systemUtils.py
@@ -3,12 +3,11 @@ from datetime import datetime
 import subprocess
 import ctypes
 import winreg
-from .orionUtils import OrionUtils
 
 class SystemUtils:
 
-    def __init__(self, orion_utils_instance): 
-        self.orion = orion_utils_instance  
+    def __init__(self, orion_utils_instance):
+        self.orion = orion_utils_instance
         self.root = self.orion.get_root_dir()
 
     def run_terminal_command(self, command):
@@ -35,36 +34,53 @@ class SystemUtils:
 
             
     def set_windows_dark_mode(self, enabled: bool):
+
         #enabled (bool): True to enable dark mode, False for light mode.
 
-        if os.name != 'nt' or winreg is None:
-            print("Dark mode control is only supported on Windows.")
-            return
+        if enabled == True:
+            if os.name != 'nt' or winreg is None:
+                print("Dark mode control is only supported on Windows.")
+                return
 
-        # 0 = Dark Mode, 1 = Light Mode
-        theme_value = 0 if enabled else 1
-        theme_action = "dark" if enabled else "light"
+            try:
+                # The registry key that controls theme settings
+                key_path = r'Software\Microsoft\Windows\CurrentVersion\Themes\Personalize'
+                
+                reg_key = winreg.CreateKey(winreg.HKEY_CURRENT_USER, key_path)
+                winreg.SetValueEx(reg_key, "AppsUseLightTheme", 0, winreg.REG_DWORD, 0)
+                winreg.SetValueEx(reg_key, "SystemUsesLightTheme", 0, winreg.REG_DWORD, 0)
+                winreg.CloseKey(reg_key)
 
-        try:
-            #registry key that controls theme settings
-            key_path = r'Software\Microsoft\Windows\CurrentVersion\Themes\Personalize'
+                print(f"Windows dark mode set to: {enabled}")
             
-            reg_key = winreg.CreateKey(winreg.HKEY_CURRENT_USER, key_path)
-            
-            #Set both App and System theme values
-            winreg.SetValueEx(reg_key, "AppsUseLightTheme", 0, winreg.REG_DWORD, theme_value)
-            winreg.SetValueEx(reg_key, "SystemUsesLightTheme", 0, winreg.REG_DWORD, theme_value)
-            winreg.CloseKey(reg_key)
+            except FileNotFoundError:
+                print("Could not find the reg key for theme.")
+            except Exception as e:
+                print(f"An error occurred while changing the theme: {e}")
 
-            print(f"Windows theme set to: {theme_action}")
+            os.system("taskkill /f /im explorer.exe")
+            subprocess.Popen("explorer.exe")
+            
+        else:
+
+            try:
+                # The registry key that controls theme settings
+                key_path = r'Software\Microsoft\Windows\CurrentVersion\Themes\Personalize'
+
+                reg_key = winreg.CreateKey(winreg.HKEY_CURRENT_USER, key_path)
         
-        except FileNotFoundError:
-            print("Could not find the reg key for theme.")
-        except Exception as e:
-            print(f"An error occurred while changing the theme: {e}")
+                is_app_light = winreg.QueryValueEx(reg_key, "AppsUseLightTheme")
+                is_sys_light = winreg.QueryValueEx(reg_key, "SystemUsesLightTheme")
 
-        os.system("taskkill /f /im explorer.exe")
-        subprocess.Popen("explorer.exe")
+                if is_app_light and is_sys_light == 0:
+                    winreg.SetValueEx(reg_key, "AppsUseLightTheme", 1, winreg.REG_DWORD, 1)
+                    winreg.SetValueEx(reg_key, "SystemUsesLightTheme", 1, winreg.REG_DWORD, 1)
+
+                else:
+                    pass
+            
+            except:
+                pass
             
     def add_line_to_file(self, file_path, line_to_add):
 
@@ -97,7 +113,8 @@ class SystemUtils:
                 files_modified.append(f)
 
         return files_modified
-    
+
+
     def env_setup(self):
         
         #set maya env var
